feat: :memo: Update AyatNesia Qur'an Search project setup and dependencies docs
This commit updates the project setup and dependencies for the AyatNesia Qur'an Search project. It includes changes to the README.md file, updating the project name and table of contents. Additionally, it updates the getting started instructions, prerequisites, and installation steps. The commit also removes unnecessary sections and reorganizes the content for better readability.  # Dependencies - Update Flutter SDK to version `>=3.1.0` - Update package dependencies to the latest versions  # Getting Started - Update project clone command to use the correct repository URL - Update project directory navigation command - Update prerequisites and installation steps  These changes ensure that the project setup and dependencies are up to date and provide accurate instructions for developers to get started with the AyatNesia Qur'an Search project.